update: Rewrite the methods of the Course and CourseController

+++ src/main/java/yoot/project_demo/Service/impl/CourseServiceImpl.java
@@ -1,11 +1,19 @@
 package yoot.project_demo.Service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import yoot.project_demo.Service.CourseService;
+import yoot.project_demo.common.exception.NotFoundException;
 import yoot.project_demo.domain.entity.Course;
+import yoot.project_demo.domain.entity.Student;
+import yoot.project_demo.dto.Course.CourseResponse;
+import yoot.project_demo.dto.Course.CourseUpsertRequest;
+import yoot.project_demo.dto.student.StudentResponse;
+import yoot.project_demo.dto.student.StudentUpsertRequest;
 import yoot.project_demo.repository.CourseRepository;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,20 +21,42 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CourseServiceImpl implements CourseService {
     private final CourseRepository courseRepository;
+    private final ModelMapper mapper;
 
-    public List<Course> findAll() {
-        return courseRepository.findAll();
+    public CourseResponse map(Course course) {
+        return mapper.map(course, CourseResponse.class);
+    }
+    public List<CourseResponse> findByAll() {
+        return courseRepository.findAll().stream()
+                .map(this::map).toList();
     }
 
-    public Optional<Course> findById(Long id) {
-        return courseRepository.findById(id);
+    public Optional<CourseResponse> findById(Long id) {
+        return courseRepository.findById(id)
+                .map(this::map);
     }
 
-    public Course save(Course course) {
-        return courseRepository.save(course);
+    public CourseResponse create(CourseUpsertRequest request) {
+        Course newCourse = mapper.map(request, Course.class);
+        newCourse.setCreateAt(LocalDateTime.now());
+        newCourse.setUpdatedAt(LocalDateTime.now());
+        Course result = courseRepository.save(newCourse);
+        return map(result);
     }
 
-    public void delete(Long id) {
-        courseRepository.deleteById(id);
+    public CourseResponse update(Long id, CourseUpsertRequest request) {
+        Course newCourse = mapper.map(request, Course.class);
+        newCourse.setId(id);
+        newCourse.setUpdatedAt(LocalDateTime.now());
+        Course result = courseRepository.save(newCourse);
+        return map(result);
+    }
+
+    public void delete(Long id) throws NotFoundException {
+        if(courseRepository.existsById(id)) {
+            courseRepository.deleteById(id);
+        } else {
+            throw new NotFoundException("Delete course error");
+        }
     }
 }
